improve StorageProvider types and refactor some context hooks
