+++ projects/bp-gallery/src/components/views/picture/sidebar/picture-info/LinkedInfoField.tsx
@@ -12,9 +12,9 @@ import { HelpTooltip } from '../../../../common/HelpTooltip';
 import PictureScrollGrid from '../../../../common/picture-gallery/PictureScrollGrid';
 import ScrollContainer from '../../../../common/ScrollContainer';
 import { AuthRole, useAuth } from '../../../../provider/AuthProvider';
+import { useClipboard } from '../../../../provider/context-hooks';
 import { DialogPreset, useDialog } from '../../../../provider/DialogProvider';
 import { ShowStatsProvider } from '../../../../provider/ShowStatsProvider';
-import { useClipboard } from '../../../../provider/StorageProvider';
 import './LinkedInfoField.scss';
 import { Field } from './PictureInfo';
 import PictureInfoField from './PictureInfoField';
